Rename some files to make their names more clear.
