+++ src/index.js
@@ -1,6 +1,6 @@
 // Bot
 export {Bot, default as createBot} from './bot';
-export {SlackBot, default as createSlackBot} from './slack/bot';
+export {SlackBot, default as createSlackBot} from './slack/slack-bot';
 
 // Message handlers
 export {DelegatingMessageHandler, default as createDelegate} from './message-handler/delegate';
