+++ src/slack/slack-bot.js
